ref: replace createListItem in listItem.ts and use new deserialisers
