+++ src/server/models/listItem/listItem.ts
@@ -1,7 +1,13 @@
 import { get, merge } from "lodash";
 import pgescape from "pg-escape";
 import { prisma } from "./../db/prisma-client";
-import { AuditEvent, Prisma, Status, ListItem, ListItemEvent } from "@prisma/client";
+import {
+  AuditEvent,
+  Prisma,
+  Status,
+  ListItem,
+  ListItemEvent,
+} from "@prisma/client";
 import { logger } from "server/services/logger";
 import {
   LawyersFormWebhookData,
@@ -21,9 +27,13 @@ import { recordEvent } from "./../listItem/listItemEvent";
 import { CovidTestSupplierListItem, LawyerListItem } from "./providers";
 import { ListItemWithAddressCountry } from "./providers/types";
 import { makeAddressGeoLocationString } from "server/models/listItem/geoHelpers";
-import { getChangedAddressFields } from "./providers/helpers";
+import {
+  checkListItemExists,
+  getChangedAddressFields,
+} from "./providers/helpers";
 import { geoLocatePlaceByText } from "server/services/location";
 import { rawUpdateGeoLocation } from "server/models/helpers";
+import { listItemCreateInputFromWebhook } from "server/models/listItem/providers/listItemCreateInputFromWebhook";
 export { findIndexListItems } from "./summary";
 
 export async function findListItemsForList(list: List): Promise<ListItem[]> {
@@ -76,11 +86,9 @@ export async function findListItemsForList(list: List): Promise<ListItem[]> {
   }
 }
 
-export async function findListItemById(
-  id: string | number
-): Promise<any> {
+export async function findListItemById(id: string | number): Promise<any> {
   try {
-    const returnVal = (await prisma.listItem.findUnique({
+    const returnVal = await prisma.listItem.findUnique({
       where: { id: Number(id) },
       include: {
         address: {
@@ -101,9 +109,8 @@ export async function findListItemById(
         history: true,
         pinnedBy: true,
       },
-    }));
+    });
     return returnVal;
-
   } catch (error) {
     logger.error(`findListItemById Error ${error.message}`);
     throw new Error("Failed to approve lawyer");
@@ -138,7 +145,8 @@ export async function togglerListItemIsApproved({
         where: { id },
         data,
       }),
-      recordListItemEvent({
+      recordListItemEvent(
+        {
           eventName: isApproved ? "approve" : "disapprove",
           itemId: id,
           userId,
@@ -166,7 +174,9 @@ export async function togglerListItemIsPublished({
     throw new Error("togglerListItemIsPublished Error: userId is undefined");
   }
   const status = isPublished ? Status.PUBLISHED : Status.UNPUBLISHED;
-  const auditEvent = isPublished ? AuditEvent.PUBLISHED : AuditEvent.UNPUBLISHED;
+  const auditEvent = isPublished
+    ? AuditEvent.PUBLISHED
+    : AuditEvent.UNPUBLISHED;
 
   try {
     const [listItem] = await prisma.$transaction([
@@ -175,7 +185,7 @@ export async function togglerListItemIsPublished({
         data: {
           isApproved: true,
           isPublished,
-          status
+          status,
         },
         include: {
           address: {
@@ -185,7 +195,8 @@ export async function togglerListItemIsPublished({
           },
         },
       }),
-      recordListItemEvent({
+      recordListItemEvent(
+        {
           eventName: isPublished ? "publish" : "unpublish",
           itemId: id,
           userId,
@@ -199,7 +210,7 @@ export async function togglerListItemIsPublished({
           itemId: id,
         },
         id,
-        isPublished ? ListItemEvent.PUBLISHED : ListItemEvent.UNPUBLISHED,
+        isPublished ? ListItemEvent.PUBLISHED : ListItemEvent.UNPUBLISHED
       ),
     ]);
 
@@ -213,7 +224,7 @@ export async function togglerListItemIsPublished({
 
 export async function persistListItemChanges(
   id: number,
-  userId: User["id"],
+  userId: User["id"]
 ): Promise<ListItem> {
   if (userId === undefined) {
     throw new Error("persistListItemChanges Error: userId is undefined");
@@ -225,13 +236,13 @@ export async function persistListItemChanges(
     const listItem = await prisma.listItem.findUnique({
       where: { id },
       include: {
-        history: true
-      }
+        history: true,
+      },
     });
 
     const editEvent = listItem?.history
       .sort((a, b) => b.time.getMilliseconds() - a.time.getMilliseconds())
-      .filter(audit => audit.type === "EDITED")
+      .filter((audit) => audit.type === "EDITED")
       .pop();
 
     const eventJsonData: EventJsonData = editEvent?.jsonData as EventJsonData;
@@ -245,20 +256,22 @@ export async function persistListItemChanges(
           isApproved: true,
           isPublished: true,
           jsonData: eventJsonData?.updatedJsonData,
-        }
+        },
       }),
-      recordListItemEvent({
+      recordListItemEvent(
+        {
           eventName: "publish",
           itemId: id,
-          userId
+          userId,
         },
         auditEvent
       ),
 
-      recordEvent({
+      recordEvent(
+        {
           eventName: "publish",
           itemId: id,
-          userId
+          userId,
         },
         id,
         listItemEvent
@@ -320,16 +333,54 @@ export async function setEmailIsVerified({
 }
 
 export async function createListItem(
-  serviceType: ServiceType,
   webhookData: WebhookData
 ): Promise<ListItemWithAddressCountry> {
-  switch (serviceType) {
-    case ServiceType.lawyers:
-      return await LawyerListItem.create(webhookData as LawyersFormWebhookData);
-    case ServiceType.covidTestProviders:
-      return await CovidTestSupplierListItem.create(
-        webhookData as CovidTestSupplierFormWebhookData
-      );
+  const exists = await checkListItemExists({
+    organisationName: webhookData.organisationName,
+    countryName: webhookData.country,
+  });
+
+  const { type } = webhookData.metadata;
+
+  if (exists) {
+    throw new Error(`${type} record already exists`);
+  }
+
+  try {
+    const data = await listItemCreateInputFromWebhook(webhookData);
+
+    const listItem = await prisma.listItem.create({
+      data,
+      include: {
+        address: {
+          include: {
+            country: true,
+          },
+        },
+      },
+    });
+
+    await recordListItemEvent(
+      {
+        eventName: "edit",
+        itemId: listItem.id,
+      },
+      AuditEvent.NEW
+    );
+
+    await recordEvent(
+      {
+        eventName: "edit",
+        itemId: listItem.id,
+      },
+      listItem.id,
+      ListItemEvent.NEW
+    );
+
+    return listItem;
+  } catch (error) {
+    logger.error(`create ListItem failed ${error.message}`);
+    throw error;
   }
 }
 
@@ -338,7 +389,9 @@ type Nullable<T> = T | undefined | null;
 export async function update(
   id: ListItem["id"],
   userId: User["id"],
-  data: WebhookDataAsJsonObject<LawyersFormWebhookData> | WebhookDataAsJsonObject<CovidTestSupplierFormWebhookData>
+  data:
+    | WebhookDataAsJsonObject<LawyersFormWebhookData>
+    | WebhookDataAsJsonObject<CovidTestSupplierFormWebhookData>
 ): Promise<void> {
   const listItemResult = await prisma.listItem
     .findFirst({
@@ -393,7 +446,8 @@ export async function update(
         prisma.listItem.update(listItemPrismaQuery),
         prisma.address.update(addressPrismaQuery!),
         rawUpdateGeoLocation(...geoLocationParams!),
-        recordListItemEvent({
+        recordListItemEvent(
+          {
             eventName: "publish",
             itemId: id,
             userId,
@@ -401,20 +455,21 @@ export async function update(
           },
           AuditEvent.PUBLISHED
         ),
-        recordEvent({
+        recordEvent(
+          {
             eventName: "publish",
             itemId: id,
-            userId
+            userId,
           },
           id,
           ListItemEvent.PUBLISHED
         ),
       ]);
-
     } else {
       result = await prisma.$transaction([
         prisma.listItem.update(listItemPrismaQuery),
-        recordListItemEvent({
+        recordListItemEvent(
+          {
             eventName: "publish",
             itemId: id,
             userId,
@@ -422,10 +477,11 @@ export async function update(
           },
           AuditEvent.PUBLISHED
         ),
-        recordEvent({
+        recordEvent(
+          {
             eventName: "publish",
             itemId: id,
-            userId
+            userId,
           },
           id,
           ListItemEvent.PUBLISHED
@@ -465,10 +521,11 @@ export async function deleteListItem(
           id,
         },
       }),
-      recordListItemEvent({
+      recordListItemEvent(
+        {
           eventName: "delete",
           itemId: id,
-          userId
+          userId,
         },
         AuditEvent.DELETED
       ),
