+++ src/server/models/listItem/providers/helpers.ts
@@ -9,11 +9,12 @@ import pgescape from "pg-escape";
 import { geoPointIsValid } from "server/models/helpers";
 import { ROWS_PER_PAGE } from "server/models/listItem/pagination";
 import { prisma } from "server/models/db/prisma-client";
-import { get, startCase, toLower } from "lodash";
+import { get, startCase, toLower, trim } from "lodash";
 import { logger } from "server/services/logger";
 import {
   CovidTestSupplierFormWebhookData,
   LawyersFormWebhookData,
+  WebhookData,
 } from "server/components/formRunner";
 import { UpdatableAddressFields } from "server/models/listItem/providers/types";
 
@@ -189,21 +190,13 @@ export function getListItemContactInformation(listItem: ListItem): {
 }
 
 export function pickWebhookAddressAsAddress(
-  webhook: LawyersFormWebhookData | CovidTestSupplierFormWebhookData
+  webhook: WebhookData
 ): Partial<UpdatableAddressFields> {
-  if ("organisationDetails" in webhook) {
-    return {
-      firstLine: webhook.organisationDetails?.addressLine1,
-      secondLine: webhook.organisationDetails?.addressLine2,
-      postCode: webhook.organisationDetails?.postcode,
-      city: webhook.organisationDetails?.city,
-    };
-  }
   return {
-    firstLine: webhook?.addressLine1,
-    secondLine: webhook?.addressLine2,
-    postCode: webhook?.postcode,
-    city: webhook?.city,
+    firstLine: webhook["address.firstLine"],
+    secondLine: webhook["address.secondLine"],
+    postCode: webhook.postCode,
+    city: webhook.city,
   };
 }
 export function getChangedAddressFields(
@@ -229,4 +222,7 @@ export function getChangedAddressFields(
       ...(valueHasChanged && { [key]: webhookValue }),
     };
   }, {});
+}
+export function checkboxCSVToArray(checkboxValue: string): string[] {
+  return checkboxValue.split(",").map(trim).filter(Boolean);
 }
